% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_trans_counts.R
\name{annual_trans_counts}
\alias{annual_trans_counts}
\title{Annual transition counts}
\usage{
annual_trans_counts(data, cycle_length = 1)
}
\arguments{
\item{data}{Individual patient data with
cvs_all:
scd:
non_cvs:
time:}

\item{cycle_length}{Break points length}
}
\description{
Aggregate individual data to annual totals.
}
