% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_sim.R
\name{ce_sim}
\alias{ce_sim}
\title{Markov model cost-effectiveness simulation}
\usage{
ce_sim(
  pop,
  probs,
  inp,
  c_unit,
  e_unit,
  c_init = NA,
  e_init = NA,
  distn = TRUE,
  labels = NULL
)
}
\arguments{
\item{pop}{Number of individuals. List-matrix \code{n_init} states x time x n_sim
Simulation parameters are taken from characteristics of pop array.}

\item{probs}{Transition probabilities. List-array \code{n_init} states x states x time x sim}

\item{c_unit}{Unit costs per state. List \code{t_max}\code{n_init}}

\item{e_unit}{Health unit values per state. List \code{n_init}}

\item{c_init}{One-off initial state costs. List \code{n_init}}

\item{labels}{Names of interventions}
}
\value{
List
}
\description{
Markov model cost-effectiveness simulation
}
\seealso{
\code{\link{init_pop}}
}
