% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_trans_mat.R
\name{group_by_trans_mat}
\alias{group_by_trans_mat}
\title{Aggregate transition matrix}
\usage{
group_by_trans_mat(data, max_year = NA)
}
\arguments{
\item{data}{new state arrivals at time step
result of running `annual_trans_counts()`
e.g. data = data_rule$`FALSE`}
}
\description{
Create aggregate observed transition counts
see p.185 Table 5.2 chapter 5 in BCEA book
}
\details{
Assume 3 state model:
1) healthy 2) scd_count 3) non_scd_count
}
