% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits_by_interv.R
\name{inits_by_interv}
\alias{inits_by_interv}
\title{Start state populations}
\usage{
inits_by_interv(ipd_risk, group_names)
}
\description{
Start state populations
}
\examples{
library(dplyr)
library(reshape2)

data("ipd_risk")
inits_by_interv(
  ipd_risk,
  group_names = c("icd", "risk_over_6", "risk_over_4"))

}
