% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_tunnel.R
\name{lambda_tunnel}
\alias{lambda_tunnel}
\title{convert BUGS transition probability matrix to
to model with tunnel state}
\usage{
lambda_tunnel(lambda, n_tunnel, alpha = 0)
}
\arguments{
\item{lambda}{S by S matrix}

\item{alpha}{proportion move from hcm to icd}
}
\description{
convert BUGS transition probability matrix to
to model with tunnel state
}
