% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_unit_fns.R
\name{rc_unit}
\alias{rc_unit}
\title{Sample costs
with tunnel states}
\usage{
rc_unit(J, inp)
}
\arguments{
\item{J}{max time (year)}

\item{inp}{unit values}
}
\value{
list
}
\description{
Sample costs
with tunnel states
}
