% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_unit_fns.R
\name{re_unit}
\alias{re_unit}
\title{sample state health
with tunnel states}
\usage{
re_unit(J, inp)
}
\arguments{
\item{J}{max time (year)}
}
\value{
list
}
\description{
sample state health
with tunnel states
}
