% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_trans_mat.R
\name{trans_counts}
\alias{trans_counts}
\title{Transition counts for multistate model}
\usage{
trans_counts(rs_name, data, CYCLE = 1)
}
\value{
list of arrays
}
\description{
Transition counts for multistate model
}
